+++ lib/src/constants/variables.dart
@@ -6,7 +6,7 @@ import '../model/prediction_model.dart';
 Rx<File?> image = Rx<File?>(null);
 String mLPredictionBaseURl = 'https://aa-api-7hrpnkb23q-nw.a.run.app/predict';
 var packagePredictionList = RxList<PredictionMlModel>().obs;
-
+String imageName = "";
 
 // response from database
 RxString nameOfMeal = "".obs;
@@ -14,8 +14,8 @@ RxBool isMeal = false.obs;
 RxBool isLoading = false.obs;
 dynamic responseData;
 RxString? downloadURL = "".obs;
-RxList<Map<String, dynamic>> packageChartDataList = <Map<String, dynamic>>[].obs;
-
+RxList<Map<String, dynamic>> packageChartDataList =
+    <Map<String, dynamic>>[].obs;
 
 // final totals of the meals.
 RxDouble totalCalories = 0.0.obs;