+++ CHANGELOG.md
@@ -34,3 +34,7 @@
 * Fixed card on example
 * Fixed api calls to prediction during the loading process
 * Added a function to combine the data for the chart
+
+## 0.0.6
+
+* Added name for the image and ID of firebase document.